internal/contour: move ClusterVisitor into internal/contour
Updates #454

Move the contents of internal/cluster into internal/contour.
