+++ internal/contour/cluster_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cluster
+package contour
 
 import (
 	"reflect"
@@ -249,7 +249,7 @@ func TestClusterVisit(t *testing.T) {
 				d.Insert(o)
 			}
 			d.Recompute()
-			v := Visitor{
+			v := clusterVisitor{
 				ClusterCache: new(ClusterCache),
 				DAG:          &d,
 			}
