+++ internal/contour/endpointstranslator.go
@@ -14,14 +14,11 @@
 package contour
 
 import (
-	"strings"
-
 	"github.com/envoyproxy/go-control-plane/envoy/api/v2"
 	"github.com/envoyproxy/go-control-plane/envoy/api/v2/core"
 	"github.com/envoyproxy/go-control-plane/envoy/api/v2/endpoint"
 	"github.com/sirupsen/logrus"
 	"k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	_cache "k8s.io/client-go/tools/cache"
 )
 
@@ -118,11 +115,11 @@ func (e *EndpointsTranslator) recomputeClusterLoadAssignment(oldep, newep *v1.En
 
 			// if this endpoint's service's port has a name, then the endpoint
 			// controller will apply the name here. The name may appear once per subset.
-			name := p.Name
-			cla, ok := clas[name]
+			portname := p.Name
+			cla, ok := clas[portname]
 			if !ok {
-				cla = clusterloadassignment(servicename(newep.ObjectMeta, name))
-				clas[name] = cla
+				cla = clusterloadassignment(servicename(newep.ObjectMeta.Namespace, newep.ObjectMeta.Name, portname))
+				clas[portname] = cla
 			}
 			for _, a := range s.Addresses {
 				cla.Endpoints[0].LbEndpoints = append(cla.Endpoints[0].LbEndpoints, lbendpoint(a.IP, p.Port))
@@ -144,10 +141,10 @@ func (e *EndpointsTranslator) recomputeClusterLoadAssignment(oldep, newep *v1.En
 		for _, p := range s.Ports {
 			// if this endpoint's service's port has a name, then the endpoint
 			// controller will apply the name here. The name may appear once per subset.
-			name := p.Name
-			if _, ok := clas[name]; !ok {
+			portname := p.Name
+			if _, ok := clas[portname]; !ok {
 				// port is not present in the list added / updated, so remove it
-				e.Remove(servicename(oldep.ObjectMeta, name))
+				e.Remove(servicename(oldep.ObjectMeta.Namespace, oldep.Name, portname))
 			}
 		}
 	}
@@ -178,17 +175,4 @@ func lbendpoint(addr string, port int32) endpoint.LbEndpoint {
 			},
 		},
 	}
-}
-
-// servicename returns a fixed name for this service and portname
-func servicename(meta metav1.ObjectMeta, portname string) string {
-	sn := []string{
-		meta.Namespace,
-		meta.Name,
-		"",
-	}[:2]
-	if portname != "" {
-		sn = append(sn, portname)
-	}
-	return strings.Join(sn, "/")
 }
